+++ wtforms_alchemy/generator.py
@@ -2,14 +2,11 @@
 from collections import OrderedDict
 from wtforms import (
     BooleanField,
-    DateField,
-    DateTimeField,
-    DecimalField,
     FloatField,
-    IntegerField,
     TextAreaField,
     TextField,
 )
+from wtforms.fields import html5
 from wtforms.validators import (
     DataRequired,
     Length,
@@ -50,20 +47,20 @@ class FormGenerator(object):
     # type objects.
     TYPE_MAP = OrderedDict((
         (sa.types.UnicodeText, TextAreaField),
-        (sa.types.BigInteger, IntegerField),
-        (sa.types.SmallInteger, IntegerField),
+        (sa.types.BigInteger, html5.IntegerField),
+        (sa.types.SmallInteger, html5.IntegerField),
         (sa.types.Text, TextAreaField),
         (sa.types.Boolean, BooleanField),
-        (sa.types.Date, DateField),
-        (sa.types.DateTime, DateTimeField),
+        (sa.types.Date, html5.DateField),
+        (sa.types.DateTime, html5.DateTimeField),
         (sa.types.Enum, SelectField),
         (sa.types.Float, FloatField),
-        (sa.types.Integer, IntegerField),
-        (sa.types.Numeric, DecimalField),
+        (sa.types.Integer, html5.IntegerField),
+        (sa.types.Numeric, html5.DecimalField),
         (sa.types.String, TextField),
         (sa.types.Time, TimeField),
         (sa.types.Unicode, TextField),
-        (types.EmailType, TextField),
+        (types.EmailType, html5.EmailField),
         (types.NumberRangeType, NumberRangeField),
         (types.PhoneNumberType, PhoneNumberField),
         (types.ColorType, ColorField),